moved authentication handlers into their own file for cleanliness
