+++ config.template.json
@@ -3,9 +3,9 @@
         "name": "makermonSession",
         "secret": "abc123"
     },
-    "domain": "colamakerspace.com",
     "oauth2":{
         "clientId": "changeme.apps.googleusercontent.com",
-        "clientSecret": "changeme"
+        "clientSecret": "changeme",
+        "requiredDomain": "colamakerspace.com"
     }
 }
